# Task2/e-transport_back/swagger-spec.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/vehicle/{id}": {
            "put": {
                "operationId": "VehicleController_updateVehicle",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateVehicleDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vehicle updated successfully."
                    },
                    "404": {
                        "description": "Vehicle not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "vehicle"
                ]
            },
            "delete": {
                "operationId": "VehicleController_deleteVehicle",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle deleted successfully."
                    },
                    "404": {
                        "description": "Vehicle not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "vehicle"
                ]
            }
        },
        "/battery": {
            "get": {
                "operationId": "BatteryController_findAllBattery",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "List of all batteries returned successfully."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "battery"
                ]
            },
            "post": {
                "operationId": "BatteryController_createBattery",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BatteryDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Battery created successfully."
                    },
                    "400": {
                        "description": "Invalid input data."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "battery"
                ]
            }
        },
        "/battery/{id}": {
            "put": {
                "operationId": "BatteryController_updateBattery",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateBatteryDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Battery updated successfully."
                    },
                    "404": {
                        "description": "Battery not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "battery"
                ]
            },
            "delete": {
                "operationId": "BatteryController_deleteBattery",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Battery deleted successfully."
                    },
                    "404": {
                        "description": "Battery not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "battery"
                ]
            }
        },
...
        "/battery-vehicle": {
            "get": {
                "operationId": "BatteryVehicleController_findAllBatteryVehicle",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "List of all battery vehicles returned successfully."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "battery-vehicle"
                ]
            },
            "post": {
                "operationId": "BatteryVehicleController_createBatteryVehicle",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BatteryVehicleDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Battery vehicle created successfully."
                    },
                    "400": {
                        "description": "Invalid input data."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "battery-vehicle"
                ]
            }
        },
        "/battery-vehicle/{id}": {
            "put": {
                "operationId": "BatteryVehicleController_updateBatteryVehicle",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateBatteryVehicleDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Battery vehicle updated successfully."
                    },
                    "404": {
                        "description": "Battery vehicle not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "battery-vehicle"
                ]
            },
            "delete": {
                "operationId": "BatteryVehicleController_deleteBatteryVehicle",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Battery vehicle deleted successfully."
                    },
                    "404": {
                        "description": "Battery vehicle not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "tags": [
                    "battery-vehicle"
                ]
            }
        }
    },
    "info": {
        "title": "e-transport",
        "description": "The e-transport API description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
...
            "UpdateVehicleDto": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "FREE",
                            "INUSE",
                            "NOTAVAILABLE",
                            "BROKEN",
                            "REPAIR"
                        ]
                    },
                    "runnedDistance": {
                        "type": "number"
                    },
                    "releaseDate": {
                        "type": "string"
                    },
                    "currentLocation": {
                        "type": "string"
                    }
                },
                "required": [
                    "status"
                ]
            },
            "BatteryDto": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "INUSE",
                            "NOTINUSE",
                            "BROKEN",
                            "CHARGING",
                            "REPAIR"
                        ]
                    },
                    "chargeLevel": {
                        "type": "number"
                    },
                    "condition": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "LithiumIon",
                            "LithiumManganese",
                            "LeadAcid"
                        ]
                    },
                    "capacity": {
                        "type": "number"
                    }
                },
                "required": [
                    "status",
                    "chargeLevel",
                    "condition",
                    "type",
                    "capacity"
                ]
            },
            "UpdateBatteryDto": {
                "type": "object",
                "properties": {
                    "chargeLevel": {
                        "type": "number"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "INUSE",
                            "NOTINUSE",
                            "BROKEN",
                            "CHARGING",
                            "REPAIR"
                        ]
                    },
                    "condition": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "LithiumIon",
                            "LithiumManganese",
                            "LeadAcid"
                        ]
                    },
                    "capacity": {
                        "type": "number"
                    }
                },
                "required": [
                    "status",
                    "type"
                ]
            },
...
            "BatteryVehicleDto": {
                "type": "object",
                "properties": {
                    "vehicleId": {
                        "type": "string"
                    },
                    "batteryId": {
                        "type": "string"
                    }
                },
                "required": [
                    "vehicleId",
                    "batteryId"
                ]
            },
            "UpdateBatteryVehicleDto": {
                "type": "object",
                "properties": {
                    "vehicleId": {
                        "type": "string"
                    },
                    "batteryId": {
                        "type": "string"
                    }
                }
            }
        }
    }
}